Committing 2 items from workspace 41ae19db-f71d-471f-9ac7-ccbc2c75ce11

+++ VALECRED_DEV/5_Notebooks/Dados_Externos/CEP/NB_Modelo_Risco_Logistico.Notebook/notebook-content.py
@@ -87,7 +87,7 @@ schema = StructType([
     StructField("lat_crise", FloatType(), True),
     StructField("lon_crise", FloatType(), True),
     StructField("severidade", StringType(), True),
-    StructField("status_ativo", IntegerType(), True)
+    StructField("status_ativo", FloatType(), True)
 ])
 
 df_pontos_crise = spark.createDataFrame(data, schema)
